+++ test/metrics/heapSizeAndUsedTest.js
@@ -1,9 +1,18 @@
 'use strict';
 
-describe('heapSizeAndUsed', () => {
+const Registry = require('../../index').Registry;
+
+describe.each([
+	['Prometheus', Registry.PROMETHEUS_CONTENT_TYPE],
+	['OpenMetrics', Registry.OPENMETRICS_CONTENT_TYPE],
+])('heapSizeAndUsed with %s registry', (tag, regType) => {
 	const heapSizeAndUsed = require('../../lib/metrics/heapSizeAndUsed');
 	const globalRegistry = require('../../lib/registry').globalRegistry;
 	const memoryUsedFn = process.memoryUsage;
+
+	beforeEach(() => {
+		globalRegistry.setContentType(regType);
+	});
 
 	afterEach(() => {
 		process.memoryUsage = memoryUsedFn;