+++ test/metrics/processRequestsTest.js
@@ -1,6 +1,11 @@
 'use strict';
 
-describe('processRequests', () => {
+const Registry = require('../../index').Registry;
+
+describe.each([
+	['Prometheus', Registry.PROMETHEUS_CONTENT_TYPE],
+	['OpenMetrics', Registry.OPENMETRICS_CONTENT_TYPE],
+])('processRequests with %s registry', (tag, regType) => {
 	const register = require('../../index').register;
 	const processRequests = require('../../lib/metrics/processRequests');
 
@@ -8,11 +13,15 @@ describe('processRequests', () => {
 		register.clear();
 	});
 
+	beforeEach(() => {
+		register.setContentType(regType);
+	});
+
 	afterEach(() => {
 		register.clear();
 	});
 
-	it('should add metric to the registry', async () => {
+	it(`should add metric to the ${tag} registry`, async () => {
 		expect(await register.getMetricsAsJSON()).toHaveLength(0);
 
 		processRequests();