+++ test/metrics/versionTest.js
@@ -1,5 +1,6 @@
 'use strict';
 
+const Registry = require('../../index').Registry;
 const nodeVersion = process.version;
 const versionSegments = nodeVersion.slice(1).split('.').map(Number);
 
@@ -15,7 +16,10 @@ function expectVersionMetrics(metrics) {
 	expect(metrics[0].values[0].labels.patch).toEqual(versionSegments[2]);
 }
 
-describe('version', () => {
+describe.each([
+	['Prometheus', Registry.PROMETHEUS_CONTENT_TYPE],
+	['OpenMetrics', Registry.OPENMETRICS_CONTENT_TYPE],
+])('version with %s registry', (tag, regType) => {
 	const register = require('../../index').register;
 	const version = require('../../lib/metrics/version');
 
@@ -23,11 +27,15 @@ describe('version', () => {
 		register.clear();
 	});
 
+	beforeEach(() => {
+		register.setContentType(regType);
+	});
+
 	afterEach(() => {
 		register.clear();
 	});
 
-	it('should add metric to the registry', async () => {
+	it(`should add metric to the ${tag} registry`, async () => {
 		expect(await register.getMetricsAsJSON()).toHaveLength(0);
 		expect(typeof versionSegments[0]).toEqual('number');
 		expect(typeof versionSegments[1]).toEqual('number');
@@ -39,7 +47,7 @@ describe('version', () => {
 		expectVersionMetrics(metrics);
 	});
 
-	it('should still be present after resetting the registry #238', async () => {
+	it(`should still be present after resetting the ${tag} registry #238`, async () => {
 		const collector = version();
 		expectVersionMetrics(await register.getMetricsAsJSON());
 		register.resetMetrics();