+++ CHANGELOG.md
@@ -18,6 +18,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - Add `get` method to type definitions of metric classes
+- Support for OpenMetrics and Exemplars
 
 ## [14.1.1] - 2022-12-31